% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvr_base.R
\name{is.lvr}
\alias{is.lvr}
\title{Check if an Object is of Class 'lvr'}
\usage{
is.lvr(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
logical
}
\description{
Tests whether the given object is an instance of class \code{lvr}.
This function performs a strict class equality check.
For a more flexible check (including subclasses), consider using \code{inherits(x, "lvr")}.
}
\examples{
SWD = lvrCreate(Name = "Stray Wild Dog", Inilv = 0.9,
                Lvsef = function(x, y) {0},
                Lvoth = function(x, y) {0.5 * y})
is.lvr(SWD)

}
