% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvr_base.R
\name{lvrCreate}
\alias{lvrCreate}
\title{Create an `lvr` Object}
\usage{
lvrCreate(
  Name = character(),
  Inilv = 0,
  Lvsef = function(x, y) {
     x
 },
  Lvoth = function(x, y) {
     y
 }
)
}
\arguments{
\item{Name}{Character; the name of the person.}

\item{Inilv}{Numeric; the initial favorability (initial value of the variable representing the person's favorability).}

\item{Lvsef}{Function; a self-influence function which must include two variables. This function defines how the person's favorability
is influenced by their own state and the other person's state.}

\item{Lvoth}{Function; an other-variable influence function which must include two variables. This function defines how the person's favorability
is influenced by the other person's state.}
}
\value{
An object of class `lvr` containing:
  \itemize{
    \item \code{name}: Name of the person.
    \item \code{inilv}: Initial favorability value.
    \item \code{lvsef}: Self-influence function.
    \item \code{lvoth}: Other-variable influence function.
  }
}
\description{
This function creates an object of class `lvr` representing a person in a two-variable dynamical system.
The object contains initial values and self- and other-variable influence functions that describe the dynamics of the system.
These dynamics are modeled as a set of coupled first-order differential equations. The `lvr` object is used for
simulating the behavior of a person within the system.
}
\examples{
someone = lvrCreate(Name = "Stray Wild Dog", Inilv = 0.9,
                    Lvsef = function(x, y) {0},
                    Lvoth = function(x, y) {0.5 * y})

}
