% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvr_base.R
\name{Ops.lvr}
\alias{Ops.lvr}
\title{Ops group's binary operations for class \code{lvr}.}
\usage{
\method{Ops}{lvr}(e1, e2)
}
\arguments{
\item{e1, e2}{An object of class \code{lvr} or a numeric.}
}
\value{
An object of class \code{lvr}.
}
\description{
These binary operators perform arithmetic operations on the initial favorability of \code{lvr}.
}
\note{
At least one of e1 and e2 is numeric and the other is object of class \code{lvr}.
}
\examples{
R <- lvrCreate(Name = "Romeo", Inilv = 0.9,
               Lvsef = function(x, y) {0},
               Lvoth = function(x, y) {0.5 * y})
J <- lvrCreate(Name = "Juliet", Inilv = 0.9,
               Lvsef = function(x, y) {0},
               Lvoth = function(x, y) {0.5 * y})

R + 0.2
0.2 + R
R - 0.2
1 - R
J * 1.1
1 / J

}
