% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvr_base.R
\name{print.lvr}
\alias{print.lvr}
\title{Print method for objects of class 'lvr'.}
\usage{
\method{print}{lvr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lvr}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for objects of class 'lvr'.
}
\examples{
SWD = lvrCreate(Name = "Stray Wild Dog", Inilv = 0.9,
                Lvsef = function(x, y) {0},
                Lvoth = function(x, y) {0.5 * y})
print(SWD)

}
